requirements.txt
alembic==1.7.3
anyio==3.3.4
asgiref==3.4.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
config==0.5.1
fastapi==0.70.0
FastAPI-SQLAlchemy==0.2.1
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
h11==0.12.0
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
Mako==1.1.5
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.2
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
starlette-context==0.3.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.6.0
